fix(app): don't re-publish received theme changes
