+++ packages/app/src/providers/DataProvider.tsx
@@ -1,10 +1,24 @@
-import { PropsWithChildren } from "react";
 import {
-  PlaylistProvider,
-  PlaylistSongProvider,
-  ProfileProvider,
-  SongProvider
-} from "./useProviders";
+  createContext,
+  Dispatch,
+  PropsWithChildren,
+  SetStateAction,
+  useCallback,
+  useContext,
+  useEffect,
+  useMemo,
+  useState
+} from "react";
+import {
+  IPCAdapter,
+  Item,
+  ItemSnapshot,
+  Playlist,
+  PlaylistSong,
+  Profile,
+  Song
+} from "../preload/ipc";
+import { useAppContext } from "./AppProvider";
 
 export const DataProvider: React.FC<PropsWithChildren> = ({ children }) => {
   return (
@@ -17,3 +31,78 @@ export const DataProvider: React.FC<PropsWithChildren> = ({ children }) => {
     </ProfileProvider>
   );
 };
+
+export const [ProfileProvider, useProfileContext] = createStoreAdapter<Profile>(
+  "profile",
+  window.ipc.adapters.profiles
+);
+export const [PlaylistProvider, usePlaylistContext] =
+  createStoreAdapter<Playlist>("playlist", window.ipc.adapters.playlists);
+export const [SongProvider, useSongContext] = createStoreAdapter<Song>(
+  "song",
+  window.ipc.adapters.songs
+);
+export const [PlaylistSongProvider, usePlaylistSongContext] =
+  createStoreAdapter<PlaylistSong>(
+    "playlist_song",
+    window.ipc.adapters.playlists_songs
+  );
+
+function createStoreAdapter<T extends Item>(
+  type: T["type"],
+  adapter: IPCAdapter<T>
+) {
+  const Context = createContext<DataStore<T> | undefined>(undefined);
+
+  const Provider: React.FC<PropsWithChildren> = ({ children }) => {
+    const [state, setState] = useState<Record<string, T>>({});
+
+    const onSnapshotChange = useCallback((snapshots: ItemSnapshot<T>[]) => {
+      setState(state => {
+        const nextState = { ...state };
+
+        for (const { action, data } of snapshots) {
+          if (action === "deleted") {
+            delete nextState[data.id];
+          } else {
+            nextState[data.id] = data;
+          }
+        }
+
+        return nextState;
+      });
+    }, []);
+
+    const { setState: setAppState } = useAppContext();
+    useEffect(() => {
+      adapter.list().then(data => {
+        onSnapshotChange(data.map(data => ({ action: "created", data })));
+        setAppState(state => ({
+          ...state,
+          dataStatus: { ...state.dataStatus, [type]: "loaded" }
+        }));
+      });
+
+      return window.ipc?.subscribe("change", (_ctx, item) => {
+        if (item.data.type !== type) return;
+        onSnapshotChange([item as ItemSnapshot<T>]);
+      });
+    }, [onSnapshotChange, setAppState]);
+
+    const value = useMemo(() => ({ state, setState }), [state]);
+    return <Context.Provider value={value}>{children}</Context.Provider>;
+  };
+
+  function useDataStore() {
+    const context = useContext(Context);
+    if (!context) throw new Error("useContext must be used within a Provider");
+    return context;
+  }
+
+  return [Provider, useDataStore] as const;
+}
+
+interface DataStore<T extends Item> {
+  state: Record<string, T>;
+  setState: Dispatch<SetStateAction<Record<string, T>>>;
+}
